requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
httpx
google-generativeai
google-cloud-speech
google-cloud-texttospeech
pydub
geocoder
pydantic